fix: widen scroll_offset to usize so scrollback beyond 65535 rows is reachable
scroll_offset was u16, silently capping the transcript scroll range at
65 535 wrapped rows — the top of big sessions was unreachable via
scroll_to_top or PgUp. Widen to usize end-to-end (scroll_up/down,
page_step, anchored-scroll maintenance in note_total_rendered, the
worker-row indicator). The final Paragraph::scroll stays u16-safe
because only the intra-viewport remainder from visible_window reaches
ratatui, and it is clamped explicitly.

+++ crates/codeoid-tui/src/app.rs
@@ -1323,8 +1323,8 @@ fn autocomplete_command(state: &mut AppState) {
 /// (more, less, vim) keeps one row of overlap on each page so the
 /// reader doesn't lose their place. Falls back to a sane default
 /// before the first frame, when we don't yet know the viewport size.
-fn page_step(state: &AppState) -> u16 {
-    state.last_viewport_rows.saturating_sub(1).max(10)
+fn page_step(state: &AppState) -> usize {
+    usize::from(state.last_viewport_rows.saturating_sub(1).max(10))
 }
 
 /// Returns `true` if the app's visual state will change on the next tick —
